py_requirements.txt
google-cloud-vision
pillow
python-Levenshtein
matplotlib